Add files via upload
Aim - To predict whether or not a student will get a placement offer.
The data was collected from a CSV file. The model used was Logistic Regression. 
Accuracy - 90%
This is an elementary-level project made without the deployment phase in my beginner stage of familiarizing myself with the Machine Learning Development Life Cycle.